Edit Schedule Page
Drop down bars for edit schedule page created.  First drop down shows
classes the user currently has that can be removed.

** CLASSES ARE NOT BEING REMOVED, it is in form brackets within html
file but it is not being saved!!!!

Second bar outputs all class/section number combinations that can be
chosen which are currently in the database

**CLASSES ARE NOT BEING SAVED
Also in form brackets but no actual form to go along and save the
information being stored